requirements.txt
Flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
flask_restful
PyJWT==2.8.0
bcrypt
flask-cors==4.0.0
flask_sqlalchemy
psycopg2
pymongo==4.6.2
Werkzeug==3.0.1
python-dateutil==2.8.2
uuid==1.30